+++ lib/l10n/app_localizations_ar.dart
@@ -1110,6 +1110,9 @@ class AppLocalizationsAr extends AppLocalizations {
   String get groceriesNameSeparator => '، ';
 
   @override
+  String get labelSeparator => '، ';
+
+  @override
   String groceriesAddedCount(int count, String shown) {
     String _temp0 = intl.Intl.pluralLogic(
       count,